# openapi.yaml
openapi: 3.1.0
info:
  title: SystemGuideForge API
  version: 0.1.0
  description: API for registering local web systems and safely analyzing authenticated screens linked through read-only navigation.

servers:
  - url: http://localhost:8080/api

tags:
  - name: Projects
    description: Projects configured in the personal workspace.
  - name: Applications
    description: Local web systems registered for analysis.
  - name: Analyses
    description: Safe read-only analysis evidence for local applications.

paths:
  /projects:
    post:
      tags: [Projects]
      summary: Create a project
      operationId: createProject
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ProjectInput'
      responses:
        '201':
          description: Project created.
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '400':
          $ref: '#/components/responses/BadRequest'

  /projects/{projectId}:
    get:
      tags: [Projects]
      summary: Get a project
      operationId: getProject
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      responses:
        '200':
          description: Project found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ProjectResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /projects/{projectId}/applications:
    post:
      tags: [Applications]
      summary: Register a local web system
      operationId: createApplication
      parameters:
        - $ref: '#/components/parameters/ProjectId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ApplicationInput'
      responses:
        '201':
          description: Application registered.
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'

  /applications/{applicationId}:
    put:
      tags: [Applications]
      summary: Update a registered local web system
      operationId: updateApplication
      parameters:
        - $ref: "#/components/parameters/ApplicationId"
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ApplicationInput"
      responses:
        "200":
          description: Application updated. Credentials are never returned.
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApplicationResponse"
        "400":
          $ref: "#/components/responses/BadRequest"
        "404":
          $ref: "#/components/responses/NotFound"
    get:
      tags: [Applications]
      summary: Get a registered application
      operationId: getApplication
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: Application found. Credentials are never returned.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApplicationResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /applications/{applicationId}/test-access:
    post:
      tags: [Applications]
      summary: Test local access with traditional username/password authentication
      description: >-
        Opens an isolated Playwright browser context, navigates to the configured
        login URL, and attempts the configured username/password login. Credentials
        are never returned. The current MVP accepts only local targets and fails
        closed when browser automation is unavailable or the login cannot be verified.
      operationId: testApplicationAccess
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: Access test result.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AccessTestResult'
        '404':
          $ref: '#/components/responses/NotFound'

  /applications/{applicationId}/analyses:
    get:
      tags: [Analyses]
      summary: List analysis history for an application
      operationId: listApplicationAnalyses
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '200':
          description: Analyses ordered newest first by start time and id.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/AnalysisSummaryResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    post:
      tags: [Analyses]
      summary: Start a safe multi-screen analysis
      description: >-
        Starts a synchronous, read-only analysis of the authenticated initial
        screen and follows same-origin links classified SAFE up to the bounded
        crawl depth. The current MVP accepts one active analysis at a time and
        never executes discovered controls.
      operationId: startAnalysis
      parameters:
        - $ref: '#/components/parameters/ApplicationId'
      responses:
        '201':
          description: Analysis completed or failed and its result was persisted.
          headers:
            Location:
              schema:
                type: string
                format: uri-reference
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: Another analysis is already running.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /analyses/{analysisId}:
    get:
      tags: [Analyses]
      summary: Get an analysis result
      operationId: getAnalysis
      parameters:
        - $ref: '#/components/parameters/AnalysisId'
      responses:
        '200':
          description: Analysis found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AnalysisResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /analyses/{analysisId}/pages:
    get:
      tags: [Analyses]
      summary: List pages discovered by an analysis
      operationId: listAnalysisPages
      parameters:
        - $ref: '#/components/parameters/AnalysisId'
      responses:
        '200':
          description: Pages discovered by the analysis.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/PageResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /pages/{pageId}/elements:
    get:
      tags: [Analyses]
      summary: List elements detected on a page
      operationId: listPageElements
      parameters:
        - $ref: '#/components/parameters/PageId'
      responses:
        '200':
          description: Elements detected on the page.
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/ElementResponse'
        '404':
          $ref: '#/components/responses/NotFound'

  /pages/{pageId}/screenshot:
    get:
      tags: [Analyses]
      summary: Get the sanitized page screenshot
      operationId: getPageScreenshot
      parameters:
        - $ref: '#/components/parameters/PageId'
      responses:
        '200':
          description: PNG screenshot with detected sensitive fields masked.
          content:
            image/png: {}
        '404':
          $ref: '#/components/responses/NotFound'

  /analyses/{analysisId}/document:
    post:
      tags: [Analyses]
      summary: Generate or retrieve the deterministic manual document draft
      operationId: generateDocument
      parameters:
        - $ref: '#/components/parameters/AnalysisId'
      responses:
        '201':
          description: Draft generated, or the existing draft reused.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          description: The analysis is not completed.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'

  /documents/{documentId}:
    put:
      tags: [Analyses]
      summary: Update a manual document draft
      operationId: updateDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/DocumentUpdateRequest'
      responses:
        '200':
          description: Updated document draft.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '400':
          description: Invalid document title, section content, or section set.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ErrorResponse'
        '404':
          $ref: '#/components/responses/NotFound'
    get:
      tags: [Analyses]
      summary: Get a generated manual document draft
      operationId: getDocument
      parameters:
        - $ref: '#/components/parameters/DocumentId'
      responses:
        '200':
          description: Document draft found.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DocumentResponse'
        '404':
          $ref: '#/components/responses/NotFound'

components:
  parameters:
    ProjectId:
      in: path
      name: projectId
      required: true
      schema:
        type: string
        format: uuid
    ApplicationId:
      in: path
      name: applicationId
      required: true
      schema:
        type: string
        format: uuid
    AnalysisId:
      in: path
      name: analysisId
      required: true
      schema:
        type: string
        format: uuid
    PageId:
      in: path
      name: pageId
      required: true
      schema:
        type: string
        format: uuid
    DocumentId:
      in: path
      name: documentId
      required: true
      schema:
        type: string
        format: uuid

  responses:
    BadRequest:
      description: Request validation failed or target URL is outside the local MVP scope.
    NotFound:
      description: Resource not found.

  schemas:
    ProjectInput:
      type: object
      additionalProperties: false
      required: [name]
      properties:
        name:
          type: string
          minLength: 1

    ProjectResponse:
      type: object
      additionalProperties: false
      required: [id, name]
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string

    ApplicationInput:
      type: object
      additionalProperties: false
      required: [name, baseUrl, loginUrl, username, password]
      properties:
        name:
          type: string
          minLength: 1
        baseUrl:
          type: string
          format: uri
          description: Local base URL, such as http://localhost:3000.
        loginUrl:
          type: string
          format: uri
          description: Local login form URL.
        username:
          type: string
          minLength: 1
          writeOnly: true
        password:
          type: string
          minLength: 1
          format: password
          writeOnly: true

    ApplicationResponse:
      type: object
      additionalProperties: false
      required: [id, projectId, name, baseUrl, loginUrl]
      properties:
        id:
          type: string
          format: uuid
        projectId:
          type: string
          format: uuid
        name:
          type: string
        baseUrl:
          type: string
          format: uri
        loginUrl:
          type: string
          format: uri

    AccessTestResult:
      type: object
      additionalProperties: false
      required: [reachable, authenticated]
      properties:
        reachable:
          type: boolean
        authenticated:
          type: boolean
        message:
          type: string

    AnalysisResponse:
      type: object
      additionalProperties: false
      required: [id, applicationId, status, startedAt]
      properties:
        id:
          type: string
          format: uuid
        applicationId:
          type: string
          format: uuid
        status:
          type: string
          enum: [RUNNING, COMPLETED, FAILED]
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        failureMessage:
          type: string
          nullable: true
          description: Generic failure message; implementation never exposes secrets.

    AnalysisSummaryResponse:
      type: object
      additionalProperties: false
      required: [id, applicationId, status, startedAt, pageCount]
      properties:
        id:
          type: string
          format: uuid
        applicationId:
          type: string
          format: uuid
        status:
          type: string
          enum: [RUNNING, COMPLETED, FAILED]
        startedAt:
          type: string
          format: date-time
        completedAt:
          type: string
          format: date-time
          nullable: true
        failureMessage:
          type: string
          nullable: true
        pageCount:
          type: integer
          format: int64
          minimum: 0

    PageResponse:
      type: object
      additionalProperties: false
      required: [id, analysisId, url, title]
      properties:
        id:
          type: string
          format: uuid
        analysisId:
          type: string
          format: uuid
        url:
          type: string
          format: uri
        title:
          type: string

    ElementResponse:
      type: object
      additionalProperties: false
      required: [id, kind, selector, actionClassification]
      properties:
        id:
          type: string
          format: uuid
        kind:
          type: string
        selector:
          type: string
        accessibleName:
          type: string
          nullable: true
        actionClassification:
          type: string
          enum: [SAFE, MUTATING, UNKNOWN]

    DocumentUpdateRequest:
      type: object
      additionalProperties: false
      required: [title, sections]
      properties:
        title: { type: string, minLength: 1, maxLength: 255 }
        sections:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSectionUpdateRequest'

    DocumentSectionUpdateRequest:
      type: object
      additionalProperties: false
      required: [id, title, content]
      properties:
        id: { type: string, format: uuid }
        title: { type: string, minLength: 1, maxLength: 255 }
        content: { type: string, minLength: 1, maxLength: 10000 }

    DocumentResponse:
      type: object
      additionalProperties: false
      required: [id, applicationId, sourceAnalysisId, title, status, sections]
      properties:
        id: { type: string, format: uuid }
        applicationId: { type: string, format: uuid }
        sourceAnalysisId: { type: string, format: uuid }
        title: { type: string }
        status: { type: string, enum: [DRAFT] }
        sections:
          type: array
          items:
            $ref: '#/components/schemas/DocumentSectionResponse'

    DocumentSectionResponse:
      type: object
      additionalProperties: false
      required: [id, position, sourcePageId, title, content]
      properties:
        id: { type: string, format: uuid }
        position: { type: integer, minimum: 0 }
        sourcePageId: { type: string, format: uuid }
        screenshotId: { type: string, format: uuid, nullable: true }
        title: { type: string }
        content: { type: string }

    ErrorResponse:
      type: object
      additionalProperties: false
      required: [message]
      properties:
        message:
          type: string